+++ native/src/main/java/io/ballerina/stdlib/http/api/HttpInterceptorUnitCallback.java
@@ -26,8 +26,6 @@ import io.ballerina.runtime.api.values.BObject;
 import io.ballerina.stdlib.http.api.nativeimpl.ModuleUtils;
 import io.ballerina.stdlib.http.transport.message.HttpCarbonMessage;
 
-import static java.lang.System.err;
-
 /**
  * {@code HttpInterceptorUnitCallback} is the responsible for acting on notifications received from Ballerina side when
  * an interceptor service is invoked.
@@ -101,13 +99,12 @@ public class HttpInterceptorUnitCallback implements Callback {
 
     private void sendRequestToNextService() {
         ballerinaHTTPConnectorListener.onMessage(requestMessage);
-        requestCtx.addNativeData(HttpConstants.REQUEST_CONTEXT_NEXT, false);
     }
 
     private void validateResponseAndProceed(Object result) {
         int interceptorId = (int) requestCtx.getNativeData(HttpConstants.INTERCEPTOR_SERVICE_INDEX);
         requestMessage.setProperty(HttpConstants.INTERCEPTOR_SERVICE_INDEX, interceptorId);
-        BArray interceptors = (BArray) requestCtx.getNativeData(HttpConstants.HTTP_INTERCEPTORS);
+        BArray interceptors = (BArray) requestCtx.getNativeData(HttpConstants.INTERCEPTORS);
         boolean nextCalled = (boolean) requestCtx.getNativeData(HttpConstants.REQUEST_CONTEXT_NEXT);
 
         if (alreadyResponded(result)) {
@@ -134,6 +131,15 @@ public class HttpInterceptorUnitCallback implements Callback {
                 } else {
                     BError err = HttpUtil.createHttpError("next interceptor service did not match " +
                             "with the configuration", HttpErrorType.GENERIC_LISTENER_ERROR);
+                    sendFailureResponse(err);
+                }
+            } else {
+                Object targetService = requestCtx.getNativeData(HttpConstants.TARGET_SERVICE);
+                if (result.equals(targetService)) {
+                    sendRequestToNextService();
+                } else {
+                    BError err = HttpUtil.createHttpError("target service did not match with the configuration",
+                            HttpErrorType.GENERIC_LISTENER_ERROR);
                     sendFailureResponse(err);
                 }
             }